+++ test/evaluacion_test.go
@@ -1,4 +1,4 @@
-package tests
+package test
 
 import (
 	"net/http"
